requirements.txt
Flask==2.2.5
requests==2.28.2
PyJWT==2.6.0 